[StyleCop] Reordering using in several classes (#1232)
* Fix all StyleCop warnings in EnglishCommonDateTimeParserConfiguration

* Fix all StyleCop warnings in BaseTimeZoneExtractor

* Fix all StyleCop warnings in IDateExtractorConfiguration

* Fix all StyleCop warnings in IDateTimeExtractor

* Fix all StyleCop warnings in IDateTimeListExtractor

* Fix all StyleCop warnings in IDateTimeZoneExtractor

* Fix all StyleCop warnings in IDurationExtractorConfiguration

* Fix StyleCop warnings in DummyTimeZoneParser

* Fix StyleCop warnings in TimeOfDayResolutionResult.cs, TimeZoneResolutionResult.cs and DateTimeResolutionResult.cs

* Fix StyleCop warnings in DateTimeFormatUtil

* Fix all StyleCop warnings in ItalianCommonDateTimeParserConfiguration

- Reordered using statements.
- Added spaces when needed.

* Fix all the StyleCop warnings in KoreanNumberParserConfiguration
-Reorder using statements
-Fix constructor initializer
