+++ .NET/Microsoft.Recognizers.Text.DateTime/Extractors/BaseTimeZoneExtractor.cs
@@ -1,8 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Text.RegularExpressions;
-using DateObject = System.DateTime;
-
 using Microsoft.Recognizers.Text.Utilities;
+using DateObject = System.DateTime;
 
 namespace Microsoft.Recognizers.Text.DateTime
 {
@@ -24,7 +23,6 @@ namespace Microsoft.Recognizers.Text.DateTime
 
         public List<ExtractResult> Extract(string text, DateObject reference)
         {
-
             var tokens = new List<Token>();
 
             var normalizedText = QueryProcessor.RemoveDiacritics(text);
