+++ .NET/Microsoft.Recognizers.Text.DateTime/Extractors/IDurationExtractorConfiguration.cs
@@ -1,7 +1,6 @@
-﻿using System.Text.RegularExpressions;
-
+﻿using System.Collections.Immutable;
+using System.Text.RegularExpressions;
 using Microsoft.Recognizers.Text.Number;
-using System.Collections.Immutable;
 
 namespace Microsoft.Recognizers.Text.DateTime
 {
