+++ .NET/Microsoft.Recognizers.Text.DateTime/Parsers/DummyTimeZoneParser.cs
@@ -1,4 +1,4 @@
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 
 using DateObject = System.DateTime;
