+++ .NET/Microsoft.Recognizers.Text.DateTime/Utilities/DateTimeFormatUtil1.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Globalization;
 using System.Text;
