+++ .NET/Microsoft.Recognizers.Text.DateTime/Utilities/TimeOfDayResolutionResult.cs
@@ -1,0 +1,15 @@
+﻿using System.Collections.Generic;
+
+namespace Microsoft.Recognizers.Text.DateTime
+{
+    public class TimeOfDayResolutionResult
+    {
+        public string Timex { get; set; }
+
+        public int BeginHour { get; set; }
+
+        public int EndHour { get; set; }
+
+        public int EndMin { get; set; }
+    }
+}
