+++ .NET/Microsoft.Recognizers.Text.DateTime/Utilities/TimeZoneResolutionResult.cs
@@ -1,0 +1,13 @@
+﻿using System.Collections.Generic;
+
+namespace Microsoft.Recognizers.Text.DateTime
+{
+    public class TimeZoneResolutionResult
+    {
+        public string Value { get; set; }
+
+        public int UtcOffsetMins { get; set; }
+
+        public string TimeZoneText { get; set; }
+    }
+}
